Create a common splitInHalfByRows split policy
